refactor(ideas): extract IdeaActions component and improve resource handling
    ## Why is this necessary?
    - IdeaCard and IdeaHeader contain duplicate edit/delete logic
    - User schema inconsistently uses both _id and id fields
    - Resource deletion in IdeaContent fails when resources have duplicate names/URLs
    - Code maintainability is reduced due to repeated logic across components

    ## How does it address?
    - Extract shared edit/delete logic into reusable IdeaActions component:
      * Create IdeaActions with dropdown menu for Edit/Delete options
      * Support custom onEdit and onDeleteSuccess callbacks for different contexts
      * Implement delete confirmation dialog with proper event handling
      * Add variant prop (icon/text) for future extensibility
    - Refactor IdeaCard to use IdeaActions component:
      * Replace inline dropdown menu with IdeaActions component
      * Remove duplicate delete dialog and action handlers
      * Simplify ownership check (remove _id fallback)
    - Refactor IdeaHeader to use IdeaActions component:
      * Replace inline dropdown menu with IdeaActions component
      * Pass onEdit and onDeleteSuccess callbacks for proper flow
      * Navigate to /explore after successful deletion
    - Fix resource management in IdeaContent:
      * Add unique crypto.randomUUID() to each resource in edit mode
      * Use resource.id as key instead of url-name combination
      * Fix handleRemoveResource to delete by id instead of index
      * Strip temporary id field before saving to backend
    - Clean up user schema in services/users/schema.ts:
      * Remove _id field from baseUserSchema
      * Use id field consistently across the application
