+++ features/ideas/components/IdeaActions.tsx
@@ -1,0 +1,138 @@
+'use client';
+
+import { useState } from 'react';
+import { useRouter } from 'next/navigation';
+import { toast } from 'sonner';
+import { Edit, Trash2, MoreVertical } from 'lucide-react';
+import { Button } from '@/shared/ui/button';
+import {
+  DropdownMenu,
+  DropdownMenuContent,
+  DropdownMenuItem,
+  DropdownMenuTrigger,
+} from '@/shared/ui/dropdown-menu';
+import {
+  AlertDialog,
+  AlertDialogAction,
+  AlertDialogCancel,
+  AlertDialogContent,
+  AlertDialogDescription,
+  AlertDialogFooter,
+  AlertDialogHeader,
+  AlertDialogTitle,
+} from '@/shared/ui/alert-dialog';
+import { useIdeaActions } from '@/features/ideas/hooks';
+import type { IdeaSchema } from '@/services/ideas';
+
+interface IdeaActionsProps {
+  idea: IdeaSchema;
+  onEdit?: () => void;
+  onDeleteSuccess?: () => void;
+  variant?: 'icon' | 'text';
+}
+
+/**
+ * 共用的 Idea 編輯/刪除操作元件
+ * 避免在 IdeaCard 和 IdeaHeader 中重複相同的邏輯
+ */
+export function IdeaActions({
+  idea,
+  onEdit,
+  onDeleteSuccess,
+  variant = 'icon'
+}: IdeaActionsProps) {
+  const router = useRouter();
+  const { deleteIdea, isDeleting } = useIdeaActions();
+  const [showDeleteDialog, setShowDeleteDialog] = useState(false);
+
+  const handleEdit = (e?: React.MouseEvent) => {
+    if (e) {
+      e.preventDefault();
+      e.stopPropagation();
+    }
+
+    if (onEdit) {
+      onEdit();
+    } else {
+      router.push(`/ideas/${idea.id}?edit=true`);
+    }
+  };
+
+  const handleDelete = async (e?: React.MouseEvent) => {
+    if (e) {
+      e.preventDefault();
+      e.stopPropagation();
+    }
+
+    try {
+      await deleteIdea(idea.id);
+      setShowDeleteDialog(false);
+      toast.success('想法已刪除');
+
+      if (onDeleteSuccess) {
+        onDeleteSuccess();
+      }
+    } catch (error) {
+      console.error('刪除想法失敗:', error);
+      toast.error('刪除失敗,請稍後再試');
+    }
+  };
+
+  return (
+    <>
+      <DropdownMenu>
+        <DropdownMenuTrigger asChild>
+          <Button
+            variant="ghost"
+            size="sm"
+            className="h-8 w-8 p-0 text-basic-400 hover:text-basic-600"
+            onClick={(e) => e.stopPropagation()}
+          >
+            <MoreVertical className="h-4 w-4" />
+          </Button>
+        </DropdownMenuTrigger>
+        <DropdownMenuContent align="end" className="w-40">
+          <DropdownMenuItem onClick={handleEdit}>
+            <Edit className="mr-2 h-4 w-4" />
+            編輯
+          </DropdownMenuItem>
+          <DropdownMenuItem
+            onClick={(e) => {
+              e.preventDefault();
+              e.stopPropagation();
+              setShowDeleteDialog(true);
+            }}
+            className="text-red-600 focus:text-red-600"
+          >
+            <Trash2 className="mr-2 h-4 w-4" />
+            刪除
+          </DropdownMenuItem>
+        </DropdownMenuContent>
+      </DropdownMenu>
+
+      {/* 刪除確認對話框 */}
+      <AlertDialog open={showDeleteDialog} onOpenChange={setShowDeleteDialog}>
+        <AlertDialogContent onClick={(e) => e.stopPropagation()}>
+          <AlertDialogHeader>
+            <AlertDialogTitle>確認刪除想法</AlertDialogTitle>
+            <AlertDialogDescription>
+              此操作無法復原。確定要刪除這個想法嗎?
+            </AlertDialogDescription>
+          </AlertDialogHeader>
+          <AlertDialogFooter>
+            <AlertDialogCancel onClick={(e) => e.stopPropagation()}>
+              取消
+            </AlertDialogCancel>
+            <AlertDialogAction
+              onClick={handleDelete}
+              disabled={isDeleting}
+              className="bg-red-600 hover:bg-red-700"
+            >
+              {isDeleting ? '刪除中...' : '確認刪除'}
+            </AlertDialogAction>
+          </AlertDialogFooter>
+        </AlertDialogContent>
+      </AlertDialog>
+    </>
+  );
+}
